+++ app/src/test/kotlin/org/agrfesta/sh/api/controllers/DevicesSynchronizeControllerMvcSliceTest.kt
@@ -10,6 +10,7 @@ import io.kotest.matchers.shouldBe
 import io.mockk.every
 import org.agrfesta.sh.api.core.application.ports.inbounds.GetDeviceUseCase
 import org.agrfesta.sh.api.core.application.ports.inbounds.GetDevicesUseCase
+import org.agrfesta.sh.api.core.application.ports.inbounds.InspectDeviceUseCase
 import org.agrfesta.sh.api.core.application.ports.inbounds.RefreshDevicesUseCase
 import org.agrfesta.sh.api.core.domain.devices.RefreshDevicesResult
 import org.agrfesta.sh.api.core.domain.failures.RefreshDevicesError
@@ -35,7 +36,8 @@ class DevicesSynchronizeControllerMvcSliceTest(
     @MockkBean private val refreshDevicesUseCase: RefreshDevicesUseCase,
     // Required by the @WebMvcTest(DevicesController) context but not exercised by these tests
     @Suppress("UnusedPrivateProperty") @MockkBean private val getDevicesUseCase: GetDevicesUseCase,
-    @Suppress("UnusedPrivateProperty") @MockkBean private val getDeviceUseCase: GetDeviceUseCase
+    @Suppress("UnusedPrivateProperty") @MockkBean private val getDeviceUseCase: GetDeviceUseCase,
+    @Suppress("UnusedPrivateProperty") @MockkBean private val inspectDeviceUseCase: InspectDeviceUseCase
 ) {
     private val authTestSupport = AuthTestSupport(mockMvc, objectMapper)